Sun, Jun 11, 2023, 12:09 PM +00:00

+++ percent.c
@@ -16,9 +16,9 @@
         printf("Your discounted price = %i \n",disc);
 
         int   discount=get_int("enter your discount: \n");
-        int   percent=(100-discount)/100;
-        int   final=percent*price;
-        printf("%i \n",final);
+       float  percent=(100-discount)/100;
+         float   final=percent*price;
+        printf("%f \n",final);
 
 
 }
